+++ coursemos_spring_boot/src/main/java/com/example/demo/domain/ScrapPK.java
@@ -11,5 +11,5 @@ import lombok.ToString;
 @Getter @Setter @AllArgsConstructor @ToString
 public class ScrapPK implements Serializable {
 	Long memberId;
-	Long courseId;
+	int courseId;
 }
